requirements.txt
openai
pydantic
pytesseract
pdf2image
pillow
pymupdf
python-dotenv
rapidfuzz
click
streamlit